requirements.txt
# -- Core ML & DL Libraries --
torch==2.3.0
transformers==4.41.2
accelerate==0.30.1
bitsandbytes==0.43.1
sentence-transformers==2.7.0

# -- Vector Database --
qdrant-client==1.9.0
# faiss-cpu==1.8.0 # Đã được thay thế bởi Qdrant

# -- Web Framework & Interface --
gradio==4.31.5
fastapi==0.111.0
uvicorn==0.30.1

# -- Data Processing (Multimedia) --
pydub==0.25.1
librosa==0.10.2
moviepy==1.0.3
opencv-python==4.9.0.80
Pillow==10.3.0
scikit-image==0.23.0

# -- LangChain (Optional, but good to have) --
# Mặc dù chúng ta không dùng nhiều nhưng có thể cần trong tương lai
langchain==0.2.1
langchain-core==0.2.1
langchain-community==0.2.1
langchain-text-splitters==0.2.0

# -- Document Loaders (for PDF, DOCX, etc.) --
pypdf==4.2.0
unstructured==0.14.4

# -- Configuration & Utilities --
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
loguru==0.7.2
tqdm==4.66.4
scikit-learn==1.5.0 # Cho cosine_similarity và các công cụ ML khác
numpy # Thường được cài đặt như một phụ thuộc, nhưng thêm vào cho chắc chắn
portalocker # Phụ thuộc của Qdrant on-disk
scipy # Cần cho Gradio khi xử lý audio numpy array